+++ example/lib/implicitly_examples_demo/implicitly_page.dart
@@ -56,6 +56,8 @@ class _ImplicitlyPageState extends State<ImplicitlyPage> {
                   onPressed: () {
                     Navigator.of(context).push(
                       MaterialPageRoute(builder: (context) => SelectExampleTypePage(
+                        description: "The animated version of the ClipRRect widget. \n\n"
+                        "Note: It's better to use AfAnimatedContainer.\nUse this widget if your case requires it.",
                         titleUnoptimizeVersion: "Without using AfAnimatedClipRRect",
                         onTapUnoptimizeVersion: () {
                           Navigator.of(context).pushNamed("/animated_clip_rrect_page_unoptimized");