+++ example/lib/main.dart
@@ -1,5 +1,7 @@
 import 'dart:ui';
 
+import 'package:af_animations/af_animations.dart';
+import 'package:af_animations_example/home_page.dart';
 import 'package:af_animations_example/implicitly_examples_demo/animated_clip_rrect_page_optimized.dart';
 import 'package:af_animations_example/implicitly_examples_demo/animated_clip_rrect_page_unoptimized.dart';
 import 'package:af_animations_example/implicitly_examples_demo/implicitly_page.dart';
@@ -10,9 +12,6 @@ import 'package:af_animations_example/value_examples_demo/animated_value_page_op
 import 'package:af_animations_example/value_examples_demo/animated_value_page_unoptimized.dart';
 import 'package:af_animations_example/value_examples_demo/value_page.dart';
 import 'package:flutter/material.dart';
-
-import 'package:af_animations/af_animations.dart';
-import 'package:af_animations_example/home_page.dart';
 
 
 void main() => runApp(const MyApp());
@@ -40,6 +39,18 @@ class _MyAppState extends State<MyApp> {
       child: MaterialApp(
         debugShowCheckedModeBanner: false,
         initialRoute: "/home",
+        theme: ThemeData(
+          appBarTheme: const AppBarTheme(
+            color: Colors.indigo,
+            titleTextStyle: TextStyle(color: Colors.white, fontSize: 20, fontWeight: FontWeight.w600),
+            iconTheme: IconThemeData(color: Colors.white),
+          ),
+          colorScheme: ColorScheme.fromSwatch(
+            primarySwatch: Colors.indigo,
+            backgroundColor: Colors.white,
+            accentColor: Colors.blueAccent,
+          ),
+        ),
         routes: {
           "/home" : (context) => const HomePage(),