+++ example/lib/other_widgets_examples_demo/af_builder_page_optimized.dart
@@ -40,9 +40,15 @@ class _AfBuilderPageOptimizedState extends State<AfBuilderPageOptimized> {
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
-            const Text(
-              '   You have pushed the button this many times:   ',
-            ).afShowRepaint(context),
+            Padding(
+              padding: const EdgeInsets.symmetric(horizontal: 20),
+              child: const Text(
+                '   You have pushed the button this many times:   ',
+                textAlign: TextAlign.center,
+              ).afShowRepaint(context),
+            ),
+
+            const SizedBox(height: 10).afShowRepaint(context),
 
             AfBuilder(
               id: "AfBuilder Optimized Demo",
@@ -52,6 +58,17 @@ class _AfBuilderPageOptimizedState extends State<AfBuilderPageOptimized> {
                   style: Theme.of(context).textTheme.headlineMedium,
                 ).afShowRepaint(context);
               },
+            ),
+
+            const SizedBox(height: 20).afShowRepaint(context),
+
+            Padding(
+              padding: const EdgeInsets.symmetric(horizontal: 20),
+              child: const Text(
+                " A color change is a widget rebuild, it may not seem like much, but in production applications, it's important to optimize everything. ",
+                textAlign: TextAlign.center,
+                textScaleFactor: 0.75,
+              ).afShowRepaint(context),
             ),
           ],
         ).afShowRepaint(context),