+++ lib/af_controller.dart
@@ -1,13 +1,13 @@
 part of 'af_animations.dart';
 
 
-/// {@template AfAnimationsController}
+/// {@template AfController}
 /// The alternative to AfAnimations without context, for those suffering from "Contextphobia"...
-/// The AfAnimationsController is used to set the duration, curve, and other animations
+/// The AfController is used to set the duration, curve, and other animations
 /// for a specified group of AfWidgets.
 /// 
 /// AfAnimations is easier and simpler to use in the application,
-/// but AfAnimationsController is a context-free alternative.
+/// but AfController is a context-free alternative.
 /// I assure you that each one has different use cases,
 /// and it will be up to your preference to choose between them.
 /// {@endtemplate}
@@ -15,7 +15,7 @@ part of 'af_animations.dart';
 /// If you find any errors, please let me know by starting a new discussion or PR :)
 /// https://github.com/FernandoDev007/af_animations
 /// 
-/// {@template AfAnimationsController_examples}
+/// {@template AfController_examples}
 /// ### Examples:
 /// ```dart
 /// class HomePage extends StatefulWidget {
@@ -27,7 +27,7 @@ part of 'af_animations.dart';
 /// 
 /// class _HomePageState extends State<HomePage> {
 /// 
-///   AfAnimationsController controller = AfAnimationsController();
+///   AfController controller = AfController();
 ///   int valueToAnimated = 0;
 /// 
 ///   @override
@@ -35,7 +35,7 @@ part of 'af_animations.dart';
 ///     /// You can also add it to a specific screen to apply it only screen
 ///     return Scaffold(
 ///       appBar: AppBar(),
-///       body: AfWidget(
+///       body: AfWidget( // Any widget from this library
 ///         controller: controller,
 ///         /// Identifier, it can be used in more than one AfWidget,
 ///         /// and by calling controller.update(ids: ["identifier group data"]),
@@ -54,11 +54,11 @@ part of 'af_animations.dart';
 /// ```
 /// {@endtemplate}
 /// 
-/// {@template AfAnimationsController_principalGetters}
+/// {@template AfController_principalGetters}
 /// Afterwards, in your project, you will be able to retrieve the specified values as follows:
 /// ```dart
 /// /// The controller is already defined previously
-/// AfAnimationsController controller = AfAnimationsController();
+/// AfController controller = AfController();
 /// 
 /// Duration globalDuration = controller.getDuration();
 /// Curve globalCurve = controller.getCurve();
@@ -66,11 +66,11 @@ part of 'af_animations.dart';
 /// ```
 /// {@endtemplate}
 /// 
-/// {@template AfAnimationsController_allGetters}
+/// {@template AfController_allGetters}
 /// You can also use these others:
 /// ```dart
 /// /// The controller is already defined previously
-/// AfAnimationsController controller = AfAnimationsController();
+/// AfController controller = AfController();
 /// 
 /// /// If for any reason you want to use it.
 /// controller.callOnEnd();
@@ -83,34 +83,34 @@ part of 'af_animations.dart';
 /// 
 /// These functions will work in these AfWidgets
 /// {@macro AfWidgets_all}
-class AfAnimationsController {
+class AfController {
 
-  /// {@macro AfAnimationsController}
-  /// {@macro AfAnimationsController_examples}
-  /// {@macro AfAnimationsController_principalGetters}
-  /// {@macro AfAnimationsController_allGetters}
-  AfAnimationsController({
+  /// {@macro AfController}
+  /// {@macro AfController_examples}
+  /// {@macro AfController_principalGetters}
+  /// {@macro AfController_allGetters}
+  AfController({
     this.duration,
     this.curve,
     this.onEnd,
     this.showRepaint = false,
   });
 
-  /// {@template AfAnimationsController_duration}
+  /// {@template AfController_duration}
   /// The global duration over which the AfWidgets will be animated.
   /// 
-  /// {@macro AfAnimationsController_principalGetters}
+  /// {@macro AfController_principalGetters}
   /// {@endtemplate}
   final Duration? duration;
 
-  /// {@template AfAnimationsController_curve}
+  /// {@template AfController_curve}
   /// The global curve over which the AfWidgets will be animated.
   /// 
-  /// {@macro AfAnimationsController_principalGetters}
+  /// {@macro AfController_principalGetters}
   /// {@endtemplate}
   final Curve? curve;
 
-  /// {@template AfAnimationsController_onEnd}
+  /// {@template AfController_onEnd}
   /// Called every time an animation completes.
   ///
   /// This can be useful to trigger additional actions (e.g. another animation)
@@ -118,7 +118,7 @@ class AfAnimationsController {
   /// {@endtemplate}
   final void Function()? onEnd;
 
-  /// {@template AfAnimationsController_showRepaint}
+  /// {@template AfController_showRepaint}
   /// If the showRepaint widget extension is set to true
   /// (You can use it in your current project to perform performance tests),
   /// all widget rebuilds will be visually displayed, making it easier to identify areas
@@ -126,7 +126,7 @@ class AfAnimationsController {
   /// 
   /// A border of a specific color will be displayed, which will change when a widget rebuild occurs.
   /// 
-  /// {@macro AfAnimationsController_principalGetters}
+  /// {@macro AfController_principalGetters}
   /// {@endtemplate}
   final bool showRepaint;
 
@@ -135,27 +135,27 @@ class AfAnimationsController {
   final List<_AfWidgetState> _afWidgetStates = <_AfWidgetState>[];
 
 
-  /// {@macro AfAnimationsController_principalGetters}
+  /// {@macro AfController_principalGetters}
   Duration getDuration() {
     return duration ?? AfAnimations._defaultDuration;
   }
 
-  /// {@macro AfAnimationsController_principalGetters}
+  /// {@macro AfController_principalGetters}
   Curve getCurve() {
     return curve ?? AfAnimations._defaultCurve;
   }
 
-  /// {@macro AfAnimationsController_principalGetters}
+  /// {@macro AfController_principalGetters}
   bool isShowRepaint() {
     return showRepaint;
   }
 
-  /// {@macro AfAnimationsController_allGetters}
+  /// {@macro AfController_allGetters}
   void callOnEnd() {
     onEnd?.call();
   }
 
-  /// {@macro AfAnimationsController_allGetters}
+  /// {@macro AfController_allGetters}
   void update({List<String> ids = const <String>[""]}) {
     for (_AfWidgetState afWidget in _afWidgetStates) {
       if (ids.any((id) => afWidget.id == id)) {
@@ -168,7 +168,7 @@ class AfAnimationsController {
 
 
 
-  /// {@template AfAnimationsController_subscription}
+  /// {@template AfController_subscription}
   /// Subscribe a _AfWidgetState to listen for a possible [controller.update]
   /// to update the widget and make the respective changes
   /// {@endtemplate}
@@ -177,7 +177,7 @@ class AfAnimationsController {
     _afWidgetStates.add(state);
   }
 
-  /// {@template AfAnimationsController_unsubscribe}
+  /// {@template AfController_unsubscribe}
   /// Cancel the subscription of a _AfWidgetState when it's not possible to listen
   /// for [controller.update], thus freeing up some resources
   /// {@endtemplate}